Added a Version Badge
Version Badge provides a consistent way for the GitHub community to learn about the GitHub repo associated with a particular Github repository and other documentation pages. Once the repo owner adds this badge to their README file, it will inform and link all visitors to the latest version of that repo.